// src/pages/Transfer/models/transferModels.ts
export type transfeTemPlateType = {
  TBLID: number;
  TEMPLATENAME: string;
  CLIENTCODE: string;
  ACCOUNTRCV: number;
  ACCOUNTRCVNAME: string;
  BANKRCVNAME: string;
  CLEAR_CODE: number;
  CREATEDTIME: Date;
  UPDATEDTIME: Date;
  BRANCHNAME: string;
  CITYCODE: string;
  BANKCODE: string;
  BANKBRANCHNAME: string;
  TEMPNAME: string;
  TEMPACCOUNTNAME: string;
};
export type hometransferds = {
  ACLIENTCODE: string;
  ACASHDEPOSITED: number;
  ALEDGERBALANCE: number;
  ACASHADVANCE: number;
  AADHOCCASH: number;
  AMARGINPRO: number;
  ACASHTRADING: number;
  ACASHMARGIN: number;
  ACASHUSED: number;
  ACASHFEE: number;
  ACASHTRANSFER: number;
  ADEBT: number;
  AFSAVING: number;
  TIENMAT: number;
  AVAIL_TRADINGCASH: number;
  AVAIL_STOCKVAL: number;
  AVAIL_TRANSFER: number;
  REMAININGQUOTA: number;
  AVAIL_PLK: number;
  AVAIL_FSAVING: number;
  AMARPRODEBT: number;
  ACASHHOLDINGBUY: number;
};